requirements.txt
requests~=2.32.4
python-telegram-bot~=22.3